Refactor: 체험 등록 api 컴포넌트 분리 작업
